+++ src/table/persistence/LokiTableMetadataStore.ts
@@ -815,8 +815,14 @@ export default class LokiTableMetadataStore implements ITableMetadataStore {
           inString = false;
         }
       } else if (query[i] === "(" || query[i] === ")") {
-        if (i !== 0 && query[i - 1].match(/\d/) !== null) {
-          // this is needed if query does not contain whitespace between number token and paren
+        if (
+          (i !== 0 &&
+            (query[i - 1].match(/\d/) !== null ||
+              (i >= 5 && query.slice(i - 5, i) === " true") ||
+              (i >= 6 && query.slice(i - 6, i) === " false"))) ||
+          query.substring(tokenStart, i).match(/\b[0-9]+L\b/g) != null
+        ) {
+          // this is needed if query does not contain whitespace between number token / boolean and paren
           appendToken();
         }
         i--;